PrevailingTrend/第一层模块/全球资金流向/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
requests>=2.31.0
python-dotenv>=1.0.0
yfinance>=0.2.0
akshare>=1.16.0
scikit-learn>=1.3.0
PyJWT>=2.8.0 